refactor: move study logic into feature model
