--- src/lib/study.spec.ts
@@ -1,199 +1,0 @@
-/**
- * @file Verifies the "resolveSwipeAction" contract with automated examples.
- * The examples make the expected behavior concrete with cases such as "returns the swipe action
- * for the given direction", "increments score for GoToNextCardMastered when score is
- * non-negative", "resets to 0 for GoToNextCardMastered when score is negative".
- */
-
-import type { Card } from "@/entities/card";
-import type { StudyPreferences, SwipeState } from "@/shared/config";
-
-import { expect, it, describe } from "vitest";
-import {
-  resolveSwipeAction,
-  calculateCardScore,
-  buildStudyPatch,
-  calculateNextIndex,
-  buildStudySession,
-  filterCardsForDeck,
-} from "@/lib/study";
-import { createDeck } from "@/test/factories";
-
-describe("resolveSwipeAction", () => {
-  it("returns the swipe action for the given direction", () => {
-    const controls = {
-      cardSwipeLeft: "GoBack",
-      cardSwipeRight: "GoToNextCardMastered",
-      cardSwipeUp: "GoToNextCardNotMastered",
-      cardSwipeDown: "DoNothing",
-    } as SwipeState;
-    expect(resolveSwipeAction(controls, "cardSwipeRight")).toBe("GoToNextCardMastered");
-    expect(resolveSwipeAction(controls, "cardSwipeLeft")).toBe("GoBack");
-    expect(resolveSwipeAction(controls, "cardSwipeUp")).toBe("GoToNextCardNotMastered");
-    expect(resolveSwipeAction(controls, "cardSwipeDown")).toBe("DoNothing");
-  });
-});
-
-describe("calculateCardScore", () => {
-  it("increments score for GoToNextCardMastered when score is non-negative", () => {
-    expect(calculateCardScore({ score: 0 } as Card, "GoToNextCardMastered")).toBe(1);
-    expect(calculateCardScore({ score: 3 } as Card, "GoToNextCardMastered")).toBe(4);
-  });
-
-  it("resets to 0 for GoToNextCardMastered when score is negative", () => {
-    expect(calculateCardScore({ score: -1 } as Card, "GoToNextCardMastered")).toBe(0);
-  });
-
-  it("decrements score for GoToNextCardNotMastered when score is non-positive", () => {
-    expect(calculateCardScore({ score: 0 } as Card, "GoToNextCardNotMastered")).toBe(-1);
-    expect(calculateCardScore({ score: -2 } as Card, "GoToNextCardNotMastered")).toBe(-3);
-  });
-
-  it("resets to 0 for GoToNextCardNotMastered when score is positive", () => {
-    expect(calculateCardScore({ score: 2 } as Card, "GoToNextCardNotMastered")).toBe(0);
-  });
-
-  it("applies same logic as NotMastered for GoToNextCardToggleMastered", () => {
-    expect(calculateCardScore({ score: 0 } as Card, "GoToNextCardToggleMastered")).toBe(-1);
-    expect(calculateCardScore({ score: 2 } as Card, "GoToNextCardToggleMastered")).toBe(0);
-  });
-
-  it("leaves score unchanged for navigation-only actions", () => {
-    expect(calculateCardScore({ score: 3 } as Card, "GoToNextCard")).toBe(3);
-    expect(calculateCardScore({ score: 3 } as Card, "GoToPrevCard")).toBe(3);
-    expect(calculateCardScore({ score: 3 } as Card, "GoBack")).toBe(3);
-    expect(calculateCardScore({ score: 3 } as Card, "DoNothing")).toBe(3);
-  });
-});
-
-describe("buildStudyPatch", () => {
-  const now = new Date(1999, 10, 1).getTime();
-  const card = { id: "c1", deckId: "d1", score: 0, numberOfSeen: 2 } as Card;
-
-  it("builds a patch with incremented numberOfSeen and computed score", () => {
-    const patch = buildStudyPatch(card, "GoToNextCardMastered", now);
-    expect(patch).toEqual({
-      id: "c1",
-      deckId: "d1",
-      score: 1,
-      numberOfSeen: 3,
-      lastSeenAt: now,
-    });
-  });
-
-  it("preserves score for navigation-only actions", () => {
-    const patch = buildStudyPatch(card, "GoToNextCard", now);
-    expect(patch.score).toBe(0);
-    expect(patch.numberOfSeen).toBe(3);
-  });
-});
-
-describe("calculateNextIndex", () => {
-  it("moves forward for non-prev actions", () => {
-    expect(calculateNextIndex(0, 3, "GoToNextCard")).toBe(1);
-    expect(calculateNextIndex(1, 3, "GoToNextCardMastered")).toBe(2);
-  });
-
-  it("returns -1 when advancing past the last card", () => {
-    expect(calculateNextIndex(2, 3, "GoToNextCard")).toBe(-1);
-  });
-
-  it("moves backward for GoToPrevCard", () => {
-    expect(calculateNextIndex(2, 3, "GoToPrevCard")).toBe(1);
-  });
-
-  it("returns -1 when moving before the first card", () => {
-    expect(calculateNextIndex(0, 3, "GoToPrevCard")).toBe(-1);
-  });
-});
-
-describe("buildStudySession", () => {
-  const cards = [{ id: "a" }, { id: "b" }, { id: "c" }, { id: "d" }] as Card[];
-
-  it("returns all card IDs in order when not shuffled and no max", () => {
-    const study = { shuffled: false, maxNumberOfCardsToLearn: 0 } as StudyPreferences;
-    expect(buildStudySession(cards, study)).toEqual(["a", "b", "c", "d"]);
-  });
-
-  it("respects maxNumberOfCardsToLearn", () => {
-    const study = { shuffled: false, maxNumberOfCardsToLearn: 2 } as StudyPreferences;
-    expect(buildStudySession(cards, study)).toEqual(["a", "b"]);
-  });
-
-  it("returns shuffled IDs when shuffled is true", () => {
-    const study = { shuffled: true, maxNumberOfCardsToLearn: 0 } as StudyPreferences;
-    const result = buildStudySession(cards, study);
-    expect(result).toHaveLength(4);
-    expect(result.sort()).toEqual(["a", "b", "c", "d"]);
-  });
-
-  it("applies max limit after shuffle", () => {
-    const study = { shuffled: true, maxNumberOfCardsToLearn: 2 } as StudyPreferences;
-    const result = buildStudySession(cards, study);
-    expect(result).toHaveLength(2);
-  });
-});
-
-describe("filterCardsForDeck", () => {
-  const now = 1_000;
-  const makeCard = (overrides: Partial<Card>): Card =>
-    ({
-      id: "c1",
-      score: 0,
-      numberOfSeen: 0,
-      tags: [],
-      nextSeeingAt: undefined,
-      ...overrides,
-    }) as Card;
-
-  const baseDeck = createDeck({
-    selectedTags: [],
-    tagAndFilter: false,
-    scoreMax: null,
-    scoreMin: null,
-  });
-
-  const baseConfig = { useCardInterval: false } as StudyPreferences;
-
-  it("returns all cards when no filters active", () => {
-    const cards = [makeCard({ id: "a" }), makeCard({ id: "b" })];
-    expect(filterCardsForDeck(cards, baseDeck, baseConfig, now)).toHaveLength(2);
-  });
-
-  it("filters by tag (OR mode)", () => {
-    const cards = [makeCard({ id: "a", tags: ["x"] }), makeCard({ id: "b", tags: ["y"] })];
-    const deck = { ...baseDeck, selectedTags: ["x"], tagAndFilter: false };
-    expect(filterCardsForDeck(cards, deck, baseConfig, now).map((c) => c.id)).toEqual(["a"]);
-  });
-
-  it("filters by tag (AND mode)", () => {
-    const cards = [makeCard({ id: "a", tags: ["x", "y"] }), makeCard({ id: "b", tags: ["x"] })];
-    const deck = { ...baseDeck, selectedTags: ["x", "y"], tagAndFilter: true };
-    expect(filterCardsForDeck(cards, deck, baseConfig, now).map((c) => c.id)).toEqual(["a"]);
-  });
-
-  it("filters by scoreMax", () => {
-    const cards = [makeCard({ id: "a", score: 3 }), makeCard({ id: "b", score: 1 })];
-    const deck = { ...baseDeck, scoreMax: 2 };
-    expect(filterCardsForDeck(cards, deck, baseConfig, now).map((c) => c.id)).toEqual(["b"]);
-  });
-
-  it("filters by scoreMin", () => {
-    const cards = [makeCard({ id: "a", score: 1 }), makeCard({ id: "b", score: 3 })];
-    const deck = { ...baseDeck, scoreMin: 2 };
-    expect(filterCardsForDeck(cards, deck, baseConfig, now).map((c) => c.id)).toEqual(["b"]);
-  });
-
-  it("filters by card interval when useCardInterval is true", () => {
-    const future = new Date(now + 100_000);
-    const cards = [makeCard({ id: "a", nextSeeingAt: future }), makeCard({ id: "b" })];
-    const config = { useCardInterval: true } as StudyPreferences;
-    expect(filterCardsForDeck(cards, baseDeck, config, now).map((c) => c.id)).toEqual(["b"]);
-  });
-
-  it("sorts by numberOfSeen ascending", () => {
-    const cards = [makeCard({ id: "a", numberOfSeen: 5 }), makeCard({ id: "b", numberOfSeen: 1 })];
-    const result = filterCardsForDeck(cards, baseDeck, baseConfig, now);
-    expect(result.map((c) => c.id)).toEqual(["b", "a"]);
-  });
-});
